+++ src/query/expression/src/lib.rs
@@ -43,7 +43,6 @@
 #![feature(debug_closure_helpers)]
 #![feature(never_type)]
 #![feature(slice_as_array)]
-#![feature(test)]
 
 #[allow(dead_code)]
 mod block;
