+++ src/query/service/src/pipelines/processors/transforms/aggregator/transform_aggregate_final.rs
@@ -16,6 +16,7 @@ use std::sync::Arc;
 
 use bumpalo::Bump;
 use databend_common_exception::Result;
+use databend_common_expression::AggregateHashTable;
 use databend_common_expression::DataBlock;
 use databend_common_expression::HashTableConfig;
 use databend_common_expression::PayloadFlushState;
@@ -27,8 +28,6 @@ use databend_common_pipeline_transforms::processors::BlockMetaTransformer;
 
 use crate::pipelines::processors::transforms::aggregator::aggregate_meta::AggregateMeta;
 use crate::pipelines::processors::transforms::aggregator::AggregatorParams;
-
-type AggregateHashTable = databend_common_expression::AggregateHashTable<false>;
 
 pub struct TransformFinalAggregate {
     params: Arc<AggregatorParams>,
