+++ src/query/service/src/pipelines/processors/transforms/aggregator/transform_aggregate_partial.rs
@@ -42,7 +42,7 @@ use crate::sessions::QueryContext;
 #[allow(clippy::enum_variant_names)]
 enum HashTable {
     MovedOut,
-    AggregateHashTable(AggregateHashTable<true>),
+    AggregateHashTable(AggregateHashTable),
 }
 
 impl Default for HashTable {
